Added a filtre option when reader .csv files
This filtre kwarg allows to filter with respect to a boolean field in
the .csv file of stations.

If in addition to 'STA', 'LAT' and 'LON', the csv files provides with a
'COND' field, then setting the filtre='cond' will allow to select only
the stations for which the 'COND' column equals 1.